refactor: RunnerPostReadRepository 삭제
