+++ backend/baton/src/main/java/touch/baton/domain/runnerpost/service/RunnerPostReadService.java
@@ -6,8 +6,9 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import touch.baton.domain.runnerpost.RunnerPost;
-import touch.baton.domain.runnerpost.repository.RunnerPostReadRepository;
+import touch.baton.domain.runnerpost.repository.RunnerPostRepository;
 import touch.baton.domain.runnerpost.vo.ReviewStatus;
+import touch.baton.domain.tag.RunnerPostTag;
 import touch.baton.domain.tag.vo.TagReducedName;
 
 import java.util.List;
@@ -17,7 +18,7 @@ import java.util.List;
 @Service
 public class RunnerPostReadService {
 
-    private final RunnerPostReadRepository runnerPostReadRepository;
+    private final RunnerPostRepository runnerPostRepository;
 
     public Page<RunnerPost> readRunnerPostByTagNameAndReviewStatus(final Pageable pageable,
                                                                    final String tagName,
@@ -25,10 +26,16 @@ public class RunnerPostReadService {
     ) {
         final TagReducedName tagReducedName = TagReducedName.from(tagName);
 
-        return runnerPostReadRepository.findByTagReducedNameAndReviewStatus(pageable, tagReducedName, reviewStatus);
+        return runnerPostRepository.findByTagReducedNameAndReviewStatus(pageable, tagReducedName, reviewStatus);
     }
 
     public List<Long> readApplicantCountsByRunnerPostIds(final List<Long> runnerPostIds) {
-        return runnerPostReadRepository.countApplicantsByRunnerPostIds(runnerPostIds);
+        return runnerPostRepository.countApplicantsByRunnerPostIds(runnerPostIds);
+    }
+
+    public List<RunnerPost> findNative(final Long cursor, final int limit, final ReviewStatus reviewStatus) {
+        final List<RunnerPost> runnerPosts = runnerPostRepository.findByPageInfoAndReviewStatus(cursor, limit, reviewStatus);
+        final List<RunnerPostTag> runnerPostTags = runnerPostRepository.findByRunnerPosts(runnerPosts);
+        return runnerPosts;
     }
 }
