+++ backend/baton/src/test/java/touch/baton/assure/repository/TestRunnerPostRepository.java
@@ -2,12 +2,24 @@ package touch.baton.assure.repository;
 
 import org.springframework.data.repository.query.Param;
 import touch.baton.domain.runnerpost.RunnerPost;
+import touch.baton.domain.runnerpost.repository.RunnerPostCustomRepository;
 import touch.baton.domain.runnerpost.repository.RunnerPostRepository;
 
-public interface TestRunnerPostRepository extends RunnerPostRepository {
+import java.util.List;
+
+public interface TestRunnerPostRepository extends RunnerPostRepository, RunnerPostCustomRepository {
 
     default RunnerPost getByRunnerPostId(@Param("runnerPostId") final Long runnerPostId) {
         return joinMemberByRunnerPostId(runnerPostId)
                 .orElseThrow(() -> new IllegalArgumentException("테스트에서 RunnerPost 를 러너 게시글 식별자값(id) 로 조회할 수 없습니다."));
+    }
+
+    default Long countApplicantByRunnerPostId(final Long runnerPostId) {
+        final List<Long> foundApplicants = countApplicantsByRunnerPostIds(List.of(runnerPostId));
+        if (foundApplicants.isEmpty()) {
+            throw new IllegalArgumentException("테스트에서 러너 게시글 식별자값으로 서포터 지원자 수 조회에 실패하였습니다.");
+        }
+
+        return foundApplicants.get(0);
     }
 }
