+++ backend/baton/src/test/java/touch/baton/config/RepositoryTestConfig.java
@@ -3,7 +3,7 @@ package touch.baton.config;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 
-@Import(JpaConfig.class)
+@Import({JpaConfig.class, QuerydslConfig.class})
 @DataJpaTest
 public abstract class RepositoryTestConfig {
 }
