+++ backend/baton/src/test/java/touch/baton/config/ServiceTestConfig.java
@@ -4,7 +4,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import touch.baton.domain.feedback.repository.SupporterFeedbackRepository;
 import touch.baton.domain.member.repository.MemberRepository;
 import touch.baton.domain.runner.repository.RunnerRepository;
-import touch.baton.domain.runnerpost.repository.RunnerPostReadRepository;
 import touch.baton.domain.runnerpost.repository.RunnerPostRepository;
 import touch.baton.domain.supporter.repository.SupporterRepository;
 import touch.baton.domain.supporter.repository.SupporterRunnerPostRepository;
@@ -27,9 +26,6 @@ public abstract class ServiceTestConfig extends RepositoryTestConfig {
 
     @Autowired
     protected RunnerPostRepository runnerPostRepository;
-
-    @Autowired
-    protected RunnerPostReadRepository runnerPostReadRepository;
 
     @Autowired
     protected SupporterRunnerPostRepository supporterRunnerPostRepository;
